Changed GetOutput Boss interaction.
Previously, before the output of a job was retrieved, the status of the job
was checked resulting in a DB query for a single job. This was done to only call
boss getOutput for jobs which are ready to retrieve the output.

Boss itself does the same check so making the first status check unnecessary.

The getoutput procedure was changed to immediately use boss getOutput and parse
the output from this command to tell the user what is going on. Three categories
are covered:

not retrieved yet
retrieving
has already been retrieved

This makes the status check before obsolete.