Provide check for default.tgz tarball size to limit input sandbox size for all submission systems.
Condor_G has no sandbox size check as EGEE does, so a size check on the
default.tgz tarball which now also contains the data directories was introduced.

The default is 100 MB.

It can be changed by:

[EDG]

maxtarballsize = X

where X is the size in MB (float).

The switch is not documented in the template crab.cfg not to mislead users.